chore: Move compression tooling into separate package
Compression tooling has been part of the chunkenc (chunk encoding)
package in the past for legacy reasons.

Since more components use this now, it's easier to keep it in a separate
package. This also eliminates the confusion around "encoding", since
this has been incorrectly used synonymously with "compression" in the past.
