+++ pkg/storage/bloom/v1/test_util.go
@@ -9,7 +9,7 @@ import (
 	"github.com/prometheus/common/model"
 	"github.com/stretchr/testify/require"
 
-	"github.com/grafana/loki/v3/pkg/chunkenc"
+	"github.com/grafana/loki/v3/pkg/compression"
 	iter "github.com/grafana/loki/v3/pkg/iter/v2"
 
 	"github.com/grafana/loki/pkg/push"
@@ -30,7 +30,7 @@ func MakeBlock(t testing.TB, nth int, fromFp, throughFp model.Fingerprint, fromT
 		BlockOptions{
 			Schema: Schema{
 				version:     CurrentSchemaVersion,
-				encoding:    chunkenc.EncSnappy,
+				encoding:    compression.EncSnappy,
 				nGramLength: 4, // see DefaultNGramLength in bloom_tokenizer_test.go
 				nGramSkip:   0, // see DefaultNGramSkip in bloom_tokenizer_test.go
 			},
