+++ pkg/storage/stores/shipper/indexshipper/uploads/index_set.go
@@ -11,7 +11,7 @@ import (
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 
-	"github.com/grafana/loki/v3/pkg/chunkenc"
+	"github.com/grafana/loki/v3/pkg/compression"
 	"github.com/grafana/loki/v3/pkg/storage/stores/shipper/indexshipper/index"
 	"github.com/grafana/loki/v3/pkg/storage/stores/shipper/indexshipper/storage"
 	util_log "github.com/grafana/loki/v3/pkg/util/log"
@@ -145,8 +145,8 @@ func (t *indexSet) uploadIndex(ctx context.Context, idx index.Index) error {
 		}
 	}()
 
-	compressedWriter := chunkenc.Gzip.GetWriter(f)
-	defer chunkenc.Gzip.PutWriter(compressedWriter)
+	compressedWriter := compression.Gzip.GetWriter(f)
+	defer compression.Gzip.PutWriter(compressedWriter)
 
 	idxReader, err := idx.Reader()
 	if err != nil {
